requirements.txt
pillow==10.3.0
tk==0.1.0
customtkinter==5.2.2
pyserial==3.5